More 64 bit warning fixes - fixes the building list display.

+++ SpotSeeker/MoreViewController.m
@@ -79,7 +79,7 @@
     }
     [super viewWillDisappear:animated];
  
-    int row = [self.campusPicker selectedRowInComponent:0];
+    NSInteger row = [self.campusPicker selectedRowInComponent:0];
     Campus *campus = [[Campus getCampuses] objectAtIndex:row];
     Campus *current_campus = [Campus getCurrentCampus];
